Run lease deadline regressions in the MySQL feature selection
